Move lease logic out of lease server.
