+++ internal/armada/server/lease.go
@@ -2,9 +2,6 @@ package server
 
 import (
 	"context"
-	"math"
-	"math/rand"
-	"time"
 
 	"github.com/gogo/protobuf/types"
 	log "github.com/sirupsen/logrus"
@@ -19,8 +16,6 @@ import (
 	"github.com/G-Research/armada/internal/armada/scheduling"
 	"github.com/G-Research/armada/internal/common"
 )
-
-const maxJobsPerLease = 10000
 
 type AggregatedQueueServer struct {
 	permissions      authorization.PermissionChecker
@@ -46,15 +41,6 @@ func NewAggregatedQueueServer(
 		queueRepository:  queueRepository,
 		usageRepository:  usageRepository,
 		eventRepository:  eventRepository}
-}
-
-type leaseContext struct {
-	ctx              context.Context
-	request          *api.LeaseRequest
-	slices           map[*api.Queue]common.ComputeResourcesFloat
-	resourceScarcity map[string]float64
-	priorities       map[*api.Queue]scheduling.QueuePriorityInfo
-	queueCache       map[string][]*api.Job
 }
 
 func (q AggregatedQueueServer) LeaseJobs(ctx context.Context, request *api.LeaseRequest) (*api.JobLease, error) {
@@ -88,45 +74,20 @@ func (q AggregatedQueueServer) LeaseJobs(ctx context.Context, request *api.Lease
 		return nil, e
 	}
 
-	resourcesToSchedule := common.ComputeResources(request.Resources).AsFloat()
-	currentClusterReport, ok := activeClusterReports[request.ClusterId]
-	if ok {
-		capacity := common.ComputeResources(currentClusterReport.ClusterCapacity)
-		resourcesToSchedule = resourcesToSchedule.LimitWith(capacity.MulByResource(q.schedulingConfig.MaximalClusterFractionToSchedule))
-	}
+	jobs, e := scheduling.LeaseJobs(
+		ctx,
+		&q.schedulingConfig,
+		q.jobRepository,
+		func(jobs []*api.Job) { reportJobsLeased(q.eventRepository, jobs, request.ClusterId) },
+		request,
+		activeClusterReports,
+		clusterPriorities,
+		queues,
+		activeQueues)
 
-	queuePriority := scheduling.CalculateQueuesPriorityInfo(clusterPriorities, activeClusterReports, queues)
-	scarcity := scheduling.ResourceScarcityFromReports(activeClusterReports)
-	activeQueuePriority := filterPriorityMapByKeys(queuePriority, activeQueues)
-	slices := scheduling.SliceResource(scarcity, activeQueuePriority, resourcesToSchedule)
-
-	jobs := []*api.Job{}
-	limit := maxJobsPerLease
-
-	lc := &leaseContext{
-		ctx:              ctx,
-		request:          request,
-		slices:           slices,
-		resourceScarcity: scarcity,
-		priorities:       activeQueuePriority,
-		queueCache:       map[string][]*api.Job{},
-	}
-
-	if !q.schedulingConfig.UseProbabilisticSchedulingForAllResources {
-		jobs, e = q.assignJobs(lc, limit)
-		if e != nil {
-			log.Errorf("Error when leasing jobs for cluster %s: %s", request.ClusterId, e)
-			return nil, e
-		}
-		limit -= len(jobs)
-	}
-
-	additionalJobs, e := q.distributeRemainder(lc, limit)
 	if e != nil {
-		log.Errorf("Error when leasing jobs for cluster %s: %s", request.ClusterId, e)
 		return nil, e
 	}
-	jobs = append(jobs, additionalJobs...)
 
 	jobLease := api.JobLease{
 		Job: jobs,
@@ -178,179 +139,4 @@ func (q *AggregatedQueueServer) ReportDone(ctx context.Context, idList *api.IdLi
 		}
 	}
 	return &api.IdList{cleanedIds}, returnedError
-}
-
-func (q *AggregatedQueueServer) assignJobs(lc *leaseContext, limit int) ([]*api.Job, error) {
-	jobs := make([]*api.Job, 0)
-	// TODO: parallelize
-	for queue, slice := range lc.slices {
-		// TODO: partition limit by priority instead
-		leased, remainder, e := q.leaseJobs(lc, queue, slice, limit/len(lc.slices))
-		if e != nil {
-			log.Error(e)
-			continue
-		}
-		lc.slices[queue] = remainder
-		jobs = append(jobs, leased...)
-
-		if closeToDeadline(lc.ctx) {
-			break
-		}
-	}
-	return jobs, nil
-}
-
-func (q *AggregatedQueueServer) distributeRemainder(lc *leaseContext, limit int) ([]*api.Job, error) {
-	jobs := []*api.Job{}
-	if limit <= 0 {
-		return jobs, nil
-	}
-
-	remainder := common.ComputeResourcesFloat{}
-	shares := map[*api.Queue]float64{}
-	for queue, slice := range lc.slices {
-		remainder.Add(slice)
-		shares[queue] = scheduling.ResourcesFloatAsUsage(lc.resourceScarcity, slice)
-	}
-
-	queueCount := len(lc.slices)
-	emptySteps := 0
-	minimumResource := q.schedulingConfig.MinimumResourceToSchedule
-	for !remainder.IsLessThan(minimumResource) && emptySteps < queueCount {
-		queue := q.pickQueueRandomly(shares)
-		emptySteps++
-
-		leased, remaining, e := q.leaseJobs(lc, queue, remainder, 1)
-		if e != nil {
-			log.Error(e)
-			continue
-		}
-		if len(leased) > 0 {
-			emptySteps = 0
-			jobs = append(jobs, leased...)
-			scheduledShare := scheduling.ResourcesFloatAsUsage(lc.resourceScarcity, remainder) - scheduling.ResourcesFloatAsUsage(lc.resourceScarcity, remaining)
-			shares[queue] = math.Max(0, shares[queue]-scheduledShare)
-			remainder = remaining
-		} else {
-			// if there are no suitable jobs to lease eliminate queue from the scheduling
-			shares[queue] = 0
-		}
-
-		limit -= len(leased)
-		if limit <= 0 || closeToDeadline(lc.ctx) {
-			break
-		}
-	}
-
-	return jobs, nil
-}
-
-func (q *AggregatedQueueServer) pickQueueRandomly(shares map[*api.Queue]float64) *api.Queue {
-	sum := 0.0
-	for _, share := range shares {
-		sum += share
-	}
-
-	pick := sum * rand.Float64()
-	current := 0.0
-
-	var lastQueue *api.Queue
-	for queue, share := range shares {
-		current += share
-		if current >= pick {
-			return queue
-		}
-		lastQueue = queue
-	}
-	log.Error("Could not randomly pick a queue, this should not happen!")
-	return lastQueue
-}
-
-func (q *AggregatedQueueServer) leaseJobs(lc *leaseContext, queue *api.Queue, slice common.ComputeResourcesFloat, limit int) ([]*api.Job, common.ComputeResourcesFloat, error) {
-	jobs := make([]*api.Job, 0)
-	remainder := slice
-	for slice.IsValid() {
-		if limit <= 0 {
-			break
-		}
-
-		topJobs, ok := lc.queueCache[queue.Name]
-		if !ok || len(topJobs) < int(q.schedulingConfig.QueueLeaseBatchSize/2) {
-			newTop, e := q.jobRepository.PeekQueue(queue.Name, int64(q.schedulingConfig.QueueLeaseBatchSize))
-			if e != nil {
-				return nil, slice, e
-			}
-			topJobs = newTop
-		}
-
-		candidates := make([]*api.Job, 0)
-		for i, job := range topJobs {
-			requirement := common.TotalResourceRequest(job.PodSpec).AsFloat()
-			remainder = slice.DeepCopy()
-			remainder.Sub(requirement)
-			if remainder.IsValid() && matchRequirements(job, lc.request) {
-				slice = remainder
-				candidates = append(candidates, job)
-
-				lc.queueCache[queue.Name] = append(topJobs[:i], topJobs[i+1:]...)
-			}
-			if len(candidates) >= limit {
-				break
-			}
-		}
-
-		leased, e := q.jobRepository.TryLeaseJobs(lc.request.ClusterId, queue.Name, candidates)
-		if e != nil {
-			return nil, slice, e
-		}
-
-		jobs = append(jobs, leased...)
-		limit -= len(leased)
-
-		// stop scheduling round if we leased less then batch (either the slice is too small or queue is empty)
-		// TODO: should we look at next batch?
-		if len(candidates) < int(q.schedulingConfig.QueueLeaseBatchSize) {
-			break
-		}
-		if closeToDeadline(lc.ctx) {
-			break
-		}
-	}
-
-	go reportJobsLeased(q.eventRepository, jobs, lc.request.ClusterId)
-
-	return jobs, slice, nil
-}
-
-func matchRequirements(job *api.Job, request *api.LeaseRequest) bool {
-	if len(job.RequiredNodeLabels) == 0 {
-		return true
-	}
-
-Labels:
-	for _, labeling := range request.AvailableLabels {
-		for k, v := range job.RequiredNodeLabels {
-			if labeling.Labels[k] != v {
-				continue Labels
-			}
-		}
-		return true
-	}
-	return false
-}
-
-func filterPriorityMapByKeys(original map[*api.Queue]scheduling.QueuePriorityInfo, keys []*api.Queue) map[*api.Queue]scheduling.QueuePriorityInfo {
-	result := make(map[*api.Queue]scheduling.QueuePriorityInfo)
-	for _, key := range keys {
-		existing, ok := original[key]
-		if ok {
-			result[key] = existing
-		}
-	}
-	return result
-}
-
-func closeToDeadline(ctx context.Context) bool {
-	d, exists := ctx.Deadline()
-	return exists && d.Before(time.Now().Add(time.Second))
 }
